Basic Go Database Implemented
A basic backend using golang was made which can store threads for the forum. Each thread currently supports a unique id, the username of whoever makes the thread, the title of the thread and the main body text. This will be modified later on based on the needs of the forum.