Frontend and Backend Communication Added
A very barebones frontend made with angular was added, which successfully communicates with the backend in order to retrieve the necessary data stored. The frontend currently allows users to enter an id, username, title, and body text for a thread and have that be posted to the backend.
The frontend server should be run by going into the frontend folder and using the command "npm run start", while the backend should be run by going into the backend folder and using the command "go run .". With both servers running simultaneously they can route data to each other thanks to a proxy that has been set up.