Massive Backend Rewrite: Database Integration
Instead of storing threads to memory, the backend now stores threads to a database called threads.db that was made using sqlite. All of the backend code was rewritten in order to query this database for information. This database integration will allow data to be stored permanently once it is added to the database, even when the backend server stops running. The database also automatically assigns an id to every new thread that is added.

Some of the new code querying the database is a bit finicky when it comes to errors, if anyone runs into any issues testing it please add an issue to the GitHub.