Added a database table for replies
I went into the current database we have and added support for replies; these replies have their own unique id, the username of who made it, the body of the actual reply, the time it was posted, and a foreign key that references the thread table in order to keep track of what replies go with what threads.
Threads and replies have a one-to-many relationship, meaning that for each thread there can be many replies, but for any reply there can only be one post.
While this database was added, backend support still needs to be added for this to be fully integrated. This includes: a struct to hold the reply data, and functions to POST and GET reply data in order to fully integrate it with the backend.