Backend Reply Functions & Go Modules Update
The backend now has support to POST replies and GET replies. This has not yet been added to gin and the frontend because the frontend page to view specific threads based on their id is not implemented yet.
The get replies function has been integrated into the getting thread by id function, which means that looking for a thread based on its ID will also access the replies table in the database to get any necessary replies.
The command "go get -u" was run to update some of the modules as well.