Backend User Authentication Added
Backend user authentication is now supported, so it can now be added to the frontend.
A new table was added to the database called “user” which stores userIDs, names and passwords. The passwords are encrypted using the bcrypt library before they’re stored to enhance security.
Four new main functions were added: register, login, currentUser and logout. Register adds the given user info to the user table and automatically assigns a unique id. Login checks if the given username and password are correct, and if they are it creates a JWT token with that information, and then creates a cookie with that token that will expire after 24 hours. currentUser checks the JWT token in the cookie to get the userID and username of the logged in user (if they are logged in, otherwise it gives an error). logout makes the cookie expire instantly, which effectively logs out the user since the login information is no longer stored. Methods to search for users in the user table by ID and name have also been added to support these purposes.
All of these functions need to be added to the documentation and unit tests need to be made for them. The user table was added to the mock database.