Unit Testing for User Authentication
Unit testing for register, login, userByName and userByID have been successfully added.
Tests for currentUser and logout were written, but since they involve cookies they did not function correctly, so the code for them was commented out. Testing for currentUser and logout has been done extensively on Postman using End-to-End testing to ensure that the cookies are the way they should be and that the two functions work properly.