Login + Register Pages Added (WIP)
The login functionality that was added to the backend has now been added to the frontend as well. Users can now register accounts and login to their account.
HOWEVER, specific user permissions and errors have not been fully implemented. This build is mainly meant to show off that the cookies sent from the backend work properly; additional UI changes, unit tests and user permissions still need to be implemented.