Added Error Handling to Login/Register
Register and Login are now able to handle any http errors that come from the backend and display messages when appropriate. Along with this, regex requirements were added to the Register page to make sure users input valid usernames and passwords. This is the first step in overhauling error handling for the entire application, and now the login/register pages are complete (besides potential UI changes and user permissions).